put back generated file

+++ include/mbgl/shaders/gl/color_relief.hpp
@@ -15,8 +15,8 @@ struct ShaderSource<BuiltIn::ColorReliefShader, gfx::Backend::Type::OpenGL> {
 layout(std140) uniform ColorReliefTilePropsUBO {
     highp vec4 u_unpack;
     highp vec2 u_dimension;
-    highp int u_color_ramp_size;
-    highp float pad_tile0;
+    int u_color_ramp_size;
+    float pad_tile0;
 };
 
 in vec2 a_pos;
@@ -41,15 +41,15 @@ void main() {
 layout(std140) uniform ColorReliefTilePropsUBO {
     highp vec4 u_unpack;
     highp vec2 u_dimension;
-    highp int u_color_ramp_size;
-    highp float pad_tile0;
+    int u_color_ramp_size;
+    float pad_tile0;
 };
 
 layout(std140) uniform ColorReliefEvaluatedPropsUBO {
-    highp float u_opacity;
-    highp float pad_eval0;
-    highp float pad_eval1;
-    highp float pad_eval2;
+    float u_opacity;
+    float pad_eval0;
+    float pad_eval1;
+    float pad_eval2;
 };
 
 uniform sampler2D u_image;
@@ -110,7 +110,8 @@ void main() {
 #ifdef OVERDRAW_INSPECTOR
     fragColor = vec4(1.0);
 #endif
-})";
+}
+)";
 };
 
 } // namespace shaders
